feat: upgrade bookSessionController to handle date and mode
This commit upgrades the bookSession function to accept the mode, datetime from req.body.
This commit also adds an updateField Object used an skeleton to propogate the non falsy values destructured from req.body.

Also this this commit now allows warden as well to use this Controller to update the session but only allows them to set mode and datetime.
Where as the client can still send datetime and mode allowing them to reschedule sessions.

And some copy changes to getSessions function